+++ packages/sdk/src/testing/util.ts
@@ -1,4 +1,6 @@
 import net from 'net';
+import { Request } from '@wundergraph/straightforward';
+import getRawBody from 'raw-body';
 
 export async function freeport(): Promise<number> {
 	const server = net.createServer();
@@ -7,4 +9,28 @@ export async function freeport(): Promise<number> {
 	const port = address.port;
 	await new Promise((resolve) => server.close(resolve));
 	return port;
+}
+
+export function getJSONBody<Body = any>(req: Request): Promise<Body> {
+	return new Promise((resolve, reject) => {
+		getRawBody(req, (err, body) => {
+			if (err) {
+				reject(err);
+			} else {
+				resolve(JSON.parse(body.toString()));
+			}
+		});
+	});
+}
+
+export function getTextBody(req: Request): Promise<string> {
+	return new Promise((resolve, reject) => {
+		getRawBody(req, (err, body) => {
+			if (err) {
+				reject(err);
+			} else {
+				resolve(body.toString());
+			}
+		});
+	});
 }