+++ pkg/hooks/pipeline.go
@@ -27,20 +27,20 @@ type QueryResolver interface {
 }
 
 type ResolveConfiguration struct {
-	// Pre indicates wheter the PreResolve hook should be run
+	// Pre indicates whether the PreResolve hook should be run
 	Pre bool
-	// MutatingPre indicates wheter the MutatingPreResolve hook should be run
+	// MutatingPre indicates whether the MutatingPreResolve hook should be run
 	MutatingPre bool
-	// Custom indicates wheter the CustomResolve hook should be run
+	// Custom indicates whether the CustomResolve hook should be run
 	Custom bool
-	// Mock indicates wether the MockResolve hook should be run
+	// Mock indicates whether the MockResolve hook should be run
 	Mock bool
 }
 
 type PostResolveConfiguration struct {
-	// Post indicates wether the PostResolve hook should be run
+	// Post indicates whether the PostResolve hook should be run
 	Post bool
-	// MutatingPost indicates wether the MutatingPostResolve hook should be run
+	// MutatingPost indicates whether the MutatingPostResolve hook should be run
 	MutatingPost bool
 }
 
@@ -130,7 +130,7 @@ func (p *pipeline) updateContextHeaders(ctx *resolve.Context, headers map[string
 		httpHeader.Set(name, headers[name])
 	}
 	ctx.Request.Header = httpHeader
-	clientRequest := ctx.Value(pool.ClientRequestKey)
+	clientRequest := ctx.Context().Value(pool.ClientRequestKey)
 	if clientRequest == nil {
 		return
 	}
@@ -173,7 +173,7 @@ func (p *pipeline) PreResolve(ctx *resolve.Context, w http.ResponseWriter, r *ht
 		if err != nil {
 			return nil, err
 		}
-		resp, err = p.client.DoOperationRequest(ctx.Context, p.operation.Name, PreResolve, hookData, payloadBuf)
+		resp, err = p.client.DoOperationRequest(ctx.Context(), p.operation.Name, PreResolve, hookData, payloadBuf)
 		if err != nil {
 			return nil, err
 		}
@@ -189,7 +189,7 @@ func (p *pipeline) PreResolve(ctx *resolve.Context, w http.ResponseWriter, r *ht
 		if err != nil {
 			return nil, err
 		}
-		resp, err = p.client.DoOperationRequest(ctx.Context, p.operation.Name, MutatingPreResolve, hookData, payloadBuf)
+		resp, err = p.client.DoOperationRequest(ctx.Context(), p.operation.Name, MutatingPreResolve, hookData, payloadBuf)
 		if err != nil {
 			return nil, err
 		}
@@ -208,7 +208,7 @@ func (p *pipeline) PreResolve(ctx *resolve.Context, w http.ResponseWriter, r *ht
 		if err != nil {
 			return nil, err
 		}
-		resp, err := p.client.DoOperationRequest(ctx.Context, p.operation.Name, MockResolve, hookData, payloadBuf)
+		resp, err := p.client.DoOperationRequest(ctx.Context(), p.operation.Name, MockResolve, hookData, payloadBuf)
 		if err != nil {
 			return nil, err
 		}
@@ -229,7 +229,7 @@ func (p *pipeline) PreResolve(ctx *resolve.Context, w http.ResponseWriter, r *ht
 		if err != nil {
 			return nil, err
 		}
-		resp, err = p.client.DoOperationRequest(ctx.Context, p.operation.Name, CustomResolve, hookData, payloadBuf)
+		resp, err = p.client.DoOperationRequest(ctx.Context(), p.operation.Name, CustomResolve, hookData, payloadBuf)
 		if err != nil {
 			return nil, err
 		}
@@ -267,7 +267,7 @@ func (p *pipeline) PostResolve(ctx *resolve.Context, w http.ResponseWriter, r *h
 		if err != nil {
 			return nil, err
 		}
-		resp, err := p.client.DoOperationRequest(ctx.Context, p.operation.Name, PostResolve, postResolveData, payloadBuf)
+		resp, err := p.client.DoOperationRequest(ctx.Context(), p.operation.Name, PostResolve, postResolveData, payloadBuf)
 		if err != nil {
 			return nil, err
 		}
@@ -283,7 +283,7 @@ func (p *pipeline) PostResolve(ctx *resolve.Context, w http.ResponseWriter, r *h
 		if err != nil {
 			return nil, err
 		}
-		resp, err := p.client.DoOperationRequest(ctx.Context, p.operation.Name, MutatingPostResolve, mutatingPostData, payloadBuf)
+		resp, err := p.client.DoOperationRequest(ctx.Context(), p.operation.Name, MutatingPostResolve, mutatingPostData, payloadBuf)
 		if err != nil {
 			return nil, err
 		}